DMA, double scanlines, service mode and cheats crash fix.
1. Use DMA for screen drawing in game.

2. Enabling cheats crashed MAME4ALL.

3. Left trigger and A button now enters 'Service Mode' (equivalent to
pressing F2 in MAME) if a game supports it. Useful for changing the
screen rotation settings in 'Dodonpachi'

4. Added 'Double Scanlines' video scaling mode. Same limitations as
Double Vertical, i.e. not really usable for games that don't originally
fit inside 320 x 240 but slightly faster as it skips a line rather than
having to duplicate to double the screen height.

5. Disabled some video options in the front end.